fix: critical bug fixes across parsing, core, and web layers (P1.3-P1.10)
P1.3: Add threading.Lock to CsvPoolService._pool_index for thread-safe
      access to the pool index dictionary. Fix false "thread-safe" docs.

P1.4: Fix scalar reduce_duplicates() using int() which silently truncates
      decimal values (e.g., 1.5 -> 1). Changed to float() to preserve
      precision during summation.

P1.5: Document that shallow copy in simple.py:184 is intentional — aliases
      share _content with parent for value aggregation. Per-file isolation
      is handled by deepcopy at line 115.

P1.6: Add bounds check in gem5_parse_work._parseLine() to prevent
      IndexError on malformed Perl output lines (< 3 parts).

P1.7: Add NaN-safe wrappers for scipy gmean/hmean in mean.py. The scipy
      functions propagate NaN unlike pandas .mean() which skips it. Now
      _safe_gmean/_safe_hmean filter NaN and handle non-positive values.

P1.8: Add None guards in mixer.py for st.segmented_control() and
      st.selectbox() returns which can be None, preventing AttributeError
      on .lower() calls.

P1.9: Add pd.isna() check alongside zero check in normalize.py for NaN
      denominators that would silently propagate through all normalized
      columns.

P1.10: Always copy DataFrame in repository_state_manager.set_data()
       before storing, preventing external mutations from propagating to
       stored data even when no type casting is needed.

+++ src/parsing/gem5/types/scalar.py
@@ -54,8 +54,8 @@ class Scalar(StatType):
             object.__setattr__(self, "_reduced_content", "NA")
             return
 
-        # Sum and divide - matching original implementation
-        total = 0
+        # Sum and divide using float to preserve decimal precision
+        total = 0.0
         for i in range(self._repeat):
-            total += int(self._content[i])
+            total += float(self._content[i])
         object.__setattr__(self, "_reduced_content", total / self._repeat)
